абстрактный класс
